# swagger.yaml
swagger: '2.0'

info:
  version: "3.0"
  title: Movie Notifier
  description: "This is a Movie Notifier server specification. Read more about the movie notifier at https://github.com/SijmenHuizenga/Movie-Notifier"
tags:
  - name: User
    description: Everything about user management and authentication.
  - name: Watcher
    description: Everything about watchers.
  - name: Operations
    description: Server status, statistics and system management.
schemes:
  - https
paths:
  /user:
    put:
      tags:
        - User
      description: Create a new user.
      parameters:
        - name: UserCreationDetails
          in: body
          description: The user data of the new user.
          required: true
          schema:
            type: object
            required:
              - name
              - email
              - phonenumber
              - password
            properties:
              name:
                title: Username
                type: string
                description: The user-friendly name of the user. The name must be between 4 and 16 charcters and can only contain letters (a-z) and numbers (0-9), but no capital letters (A-Z). The first 4 characters must always be letters.
              email:
                title: Email
                type: string
                format: email
                description: A valid email adres.
              phonenumber:
                title: Phone Number
                type: string
                description: A valid, Global Number as described in RFC 3966 section 5.1.4 (always in the format of +[countrycode][phonenumber]).
              password:
                title: Password
                type: string
                format: password
                description: The password of the user. The password must at least be 6 characters long and may only contain the letters (a-z), capital letters (A-Z), numbers (0-9) and the following special characters between (and thus except) the quotation marks "!@#$%^&*()_-+={}[]:;?><.,".
      responses:
        201:
         description: Returns the details about the created user
         schema:
           $ref: '#/definitions/User'
        400:
          $ref: '#/responses/BadRequestError'
        500:
          $ref: '#/responses/InternalServerError'
  /user/login:
    post:
      tags:
        - User
      description: Login a user.
      parameters:
        - name: UserLoginDetails
          in: body
          description: The login details
          required: true
          schema:
            type: object
            required:
              - name
              - password
            properties:
              name:
                title: Username
                type: string
                description: The user-friendly name of the user.
              password:
                title: Password
                type: string
                format: password
                description: The password of the user.

      responses:
        200:
         description: The login is successfull. The user details are returned.
         schema:
           $ref: '#/definitions/User'
        401:
          description: Not Authorized. The name/password combination is not correct.
        400:
          $ref: '#/responses/BadRequestError'
        500:
          $ref: '#responses/InternalServerError'
  /user/{userid}:
    get:
      tags:
        - User
      description: Get user details
      parameters:
        - $ref: '#/parameters/userid'
      responses:
        200:
         description: Returns the details about the user. Only the fields that you have permission to see are returned.
         schema:
           $ref: '#/definitions/User'
        400:
          $ref: '#/responses/BadRequestError'
        401:
          $ref: "#/responses/UnauthorizedError"
        500:
          $ref: '#/responses/InternalServerError'
      security:
        - APIKEY: []
    post:
      tags:
        - User
      description: Updates a user
      parameters:
        - $ref: '#/parameters/userid'
        - name: UserUpdateDetails
          in: body
          description: The fields that should be updated. Fields that do not require updates can be omitted.
          required: true
          schema:
            type: object
            properties:
              name:
                title: Username
                type: string
                description: The user-friendly name of the user. The name must be between 4 and 16 charcters and can only contain letters (a-z) and numbers (0-9), but no capital letters (A-Z). The first 4 characters must always be letters.
              email:
                title: Email
                type: string
                format: email
                description: A valid email adres.
              phonenumber:
                title: Phone Number
                type: string
                description: A valid, Global Number as described in RFC 3966 section 5.1.4 (always in the format of +[countrycode][phonenumber]).
              password:
                title: Password
                type: string
                format: password
                description: The password of the user. The password must at least be 6 characters long and may only contain the letters (a-z), capital letters (A-Z), numbers (0-9) and the following special characters between (and thus except) the quotation marks "!@#$%^&*()_-+={}[]:;?><.,".
      responses:
        200:
         description: Returns the details about the updated user
         schema:
           $ref: '#/definitions/User'
        400:
          $ref: '#/responses/BadRequestError'
        401:
          $ref: "#/responses/UnauthorizedError"
        500:
          $ref: '#/responses/InternalServerError'
      security:
        - APIKEY: []
    delete:
      tags:
        - User
      description: Deletes a user and all corresponding data in the system.
      parameters:
        - $ref: '#/parameters/userid'
      responses:
        200:
         description: The user is succesfully deleted
        400:
          $ref: '#/responses/BadRequestError'
        401:
          $ref: "#/responses/UnauthorizedError"
        500:
          $ref: '#/responses/InternalServerError'
      security:
        - APIKEY: []
  /user/{userid}/notifications/:
    get:
      tags:
        - User
      description: Get the notification settings of a user.
      parameters:
        - $ref: '#/parameters/userid'
      responses:
        200:
          description: The notification settings of the given user
          schema:
            type: array
            items:
              $ref: '#/definitions/NotificationSetting'
        400:
          description: Bad Request
        401:
          $ref: "#/responses/UnauthorizedError"
        500:
          description: Internal server error
      security:
        - APIKEY: []
    post:
      tags:
        - User
      description: Update the notification settings of a user.
      parameters:
        - $ref: '#/parameters/userid'
        - name: notificationsettings
          in: body
          required: true
          description: The new notification settings of a given user. All notification types that are not specified are set to the default value false.
          schema:
            type: array
            items:
              $ref: '#/definitions/NotificationSetting'
      responses:
        200:
          description: The updates are saved. Response are the new total notification settings.
          schema:
            type: array
            items:
              $ref: '#/definitions/NotificationSetting'
        400:
          description: Bad Request
        401:
          $ref: "#/responses/UnauthorizedError"
        500:
          description: Internal server error
      security:
        - APIKEY: []
  /watchers:
    get:
      tags:
        - Watcher
      description: Get a list of all watchers
      responses:
        200:
          description: All the watchers.
          schema:
            type: array
            items:
              $ref: '#/definitions/Watcher'
        400:
          description: Bad Request
        401:
          $ref: "#/responses/UnauthorizedError"
        500:
          description: Internal server error
      security:
        - APIKEY: []
  /watchers/{watcherid}:
    get:
      tags:
        - Watcher
      description: Get the details of the given watcher id
      parameters:
        - $ref: '#/parameters/watcherid'
      responses:
        200:
          description: Watcher details.
          schema:
            $ref: '#/definitions/Watcher'
        400:
          description: Bad Request
        401:
          $ref: "#/responses/UnauthorizedError"
        500:
          description: Internal server error
      security:
        - APIKEY: []
    post:
      tags:
        - Watcher
      description: Update the details of the given watcher id
      parameters:
        - $ref: '#/parameters/watcherid'
        - name: watcher
          in: body
          description: The watcher data to update. The fields `uuid` and `user` should not be provided as it is not possible to update these fields.
          schema:
            $ref: '#/definitions/Watcher'
          required: true
      responses:
        200:
          description: The update is succesful. Returned is the created watcher.
          schema:
            $ref: '#/definitions/Watcher'
        400:
          description: Bad Request
          schema:
            type: string
            description: The reason for creation failure.
        401:
          $ref: "#/responses/UnauthorizedError"
        500:
          description: Internal server error
      security:
        - APIKEY: []
    delete:
      tags:
        - Watcher
      description: Delete the watcher given the watcher id
      parameters:
        - $ref: '#/parameters/watcherid'
      responses:
        200:
          description: The deletion is succesful.
        400:
          description: Bad Request
        401:
          $ref: "#/responses/UnauthorizedError"
        500:
          description: Internal server error
      security:
        - APIKEY: []
  /watchers/user/{userid}:
    get:
      tags:
        - Watcher
      description: Get a list of all watchers given a user.
      parameters:
        - $ref: '#/parameters/userid'
      responses:
        200:
          description: All the watchers for the given user.
          schema:
            type: array
            items:
              $ref: '#/definitions/Watcher'
        400:
          description: Bad Request
        401:
          $ref: "#/responses/UnauthorizedError"
        500:
          description: Internal server error
      security:
        - APIKEY: []
    put:
      tags:
        - Watcher
      description: Create a new watcher
      parameters:
        - $ref: '#/parameters/userid'
        - name: watcher
          in: body
          description: The new watcher data. The field `uuid` must be omitted since this field will be created by the system. The field `user` can be ommitted. When empty the current user will be assigned to this watcher.
          schema:
            $ref: '#/definitions/Watcher'
          required: true
      responses:
        200:
          description: The creation is succesful. Returned is the created watcher.
          schema:
            $ref: '#/definitions/Watcher'
        400:
          description: Bad Request
          schema:
            type: string
            description: The reason for creation failure.
        401:
          $ref: "#/responses/UnauthorizedError"
        500:
          description: Internal server error
      security:
        - APIKEY: []
  /watchers/movies:
    get:
      tags:
        - Watcher
      description: Get a list of all movies that have watchers
      responses:
        200:
          description: All the movie ID's that contain watchers
          schema:
            type: array
            items:
              type: integer
              title: movieid
              description: The unique identifier of the movie
        400:
          description: Bad Request
        401:
          $ref: "#/responses/UnauthorizedError"
        500:
          description: Internal server error
      security:
        - APIKEY: []
  /watchers/movies/{movieid}:
    get:
      tags:
        - Watcher
      description: Get a list of all watchers that watch the given movie id
      parameters:
        - name: movieid
          in: path
          description: The unique identifier of the movie that is searched for.
          required: true
          type: integer
      responses:
        200:
          description: All the watchers for the given movie.
          schema:
            type: array
            items:
              $ref: '#/definitions/Watcher'
        400:
          description: Bad Request
        401:
          $ref: "#/responses/UnauthorizedError"
        500:
          description: Internal server error
      security:
        - APIKEY: []
  /status:
    get:
      tags:
        - Operations
      description: Ping to this server to check the health
      responses:
        200:
          description: The service is ok.
        503:
          description: The service is not ok.
          schema:
             type: string
             title: description
             description: The reason of the failure of the service
  /status/notifications:
    get:
      tags:
        - Operations
      description: Get the status of the notification service
      responses:
        200:
          description: The status
          schema:
            type: array
            title: Notification service status
            items:
              type: object
              properties:
                notificationtypeid:
                  description: A three character string that uniquely identifies a notification type.
                  type: string
                usercountenabled:
                  description: the amount of users that have this notification type enabled.
  /status/watchers:
    get:
      tags:
        - Operations
      description: Get the status of this watcher service
      responses:
        200:
          description: The status
          schema:
            type: object
            title: Watcher Service Status
            properties:
              servicestart:
                description: The moment this service was first started. All statistics are from this date.
                type: string
                format: date-time
              watchercounts:
                description: The total amount of watchers
  /status/monitor:
    get:
      tags:
        - Operations
      description: Get the status of the pathe monitor service
      responses:
        200:
          description: The status
          schema:
            type: object
            title: Monitor Service Status
            properties:
              servicestart:
                description: The moment this service was first started. All statistics are from this date.
                type: string
                format: date-time
              lastcheck:
                description: The last timestamp of a succesfull pathé api call
                type: string
                format: date-time
              lastfailedcheck:
                description: The last timestamp of a failed pathé api call
                type: string
                format: date-time
              apicallstotal:
                description: The total amount of api calls since the start of this module
                type: integer
              apicallsfailed:
                description: The total amount of failed api calls since the start of this module
                type: integer
              apicallstotaltoday:
                description: The total amount of api calls today
                type: integer
              apicallsfailedtoday:
                description: The total amount of failed api calls today
                type: integer
              apicallstotallasthour:
                description: The total amount of api calls last hour
                type: integer
              apicallsfailedlasthour:
                description: The total amount of failed api calls last hour
                type: integer
              pollrate:
                description: The amount of seconds between polls at this current moment
                type: integer
              pollrates:
                description: The schema of how often the api is polled in what time periods.
                type: array
                items:
                  type: object
                  properties:
                    start:
                      description: The moment this pollrate is activated.
                      type: string
                    end:
                      description: The moment this pollrate is stopped.
                      type: string
                    rate:
                      description: The amount of seconds between polls
                      type: integer
securityDefinitions:
  APIKEY:
    type: apiKey
    description: |
      A 64 character string that authenticates the request.
      This apikey is received by loggin in using the url /user/login.
    name: APIKEY
    in: header
parameters:
  userid:
    name: userid
    in: path
    description: The unique identifier of the user.
    required: true
    type: string
  watcherid:
    name: watcherid
    in: path
    description: The unique identifier of a watcher.
    required: true
    type: string
definitions:
  User:
    type: object
    description: The full representation of a user.
    required:
      - uuid
      - name
      - email
      - phonenumber
      - apikey
    properties:
      uuid:
        type: string
        description: A Universally Unique IDentifier.
      name:
        title: Username
        type: string
        description: The user-friendly name of the user. Is always between 4 and 16 charcters and only contains letters (a-z) and numbers (0-9), but no capital letters (A-Z). The first 4 characters are always letters.
      email:
        title: Email
        type: string
        format: email
        description: A valid email adres.
      phonenumber:
        title: Phone Number
        type: string
        description: A valid, Global Number as described in RFC 3966 section 5.1.4 (always in the format of +[countrycode][phonenumber]).
      apikey:
        description: A 64 character string that authenticates the user in api requests.
        type: string
        minLength: 64
        maxLength: 64
  NotificationSetting:
    type: object
    properties:
      notificationtypeid:
        description: A three character string that uniquely identifies a notification type.
        type: string
      name:
        description: The name of this notification type
        type: string
      description:
        description: The extended description of this notification type
        type: string
      enabled:
        description: Weather or not this notification type is enabled
        type: boolean
  Watcher:
    type: object
    description: The full representation of a watcher.
    required:
      - name
      - movieid
      - cinemaid
    properties:
      uuid:
        type: string
        description: The unique identifier that identifies this watcher. Format is a Universally Unique IDentifier as described in RFC 4122.
      user:
        type: string
        description: The unique identifier of the user that owns this watcher.
      name:
        type: string
        description: The displayname of this watcher.
      movieid:
        type: integer
        description: The unique identifier of the movie to watch for.
      cinemaid:
        type: integer
        description: The unique identifier of the movie to watch for. If empty than every cinema is acceptable.
      startafter:
        type: string
        format: date-time
        description: The ealyest moment of a showing to watch for.
      startBefore:
        type: string
        format: date-time
        description: The latest moment of a showing to watch for.
      props:
        type: object
        properties:
          imax:
            type: boolean
            description: whether or not the showing is in
          3d:
            type: boolean
            description: whether or not the showing is in 3D
          ov:
            type: boolean
            description: whether or not the showing is in the original language
          nl:
            type: boolean
            description: whether or not the showing is in the dutch language
          vip:
            type: boolean
            description: whether or not the showing is a VIP showing
          hfr:
            type: boolean
            description: whether or not the showing is in HFR
          largehall:
            type: boolean
            description: whether or not the showing is in displayed in a large hall
          atmos:
            type: boolean
            description: whether or not the showing is in dolbey atmos
          4k:
            type: boolean
            description: whether or not the showing is in 4K resolution
          laser:
            type: boolean
            description: whether or not the showing is projected with a laser projector
          premium:
            type: boolean
            description: whether or not the showing is available with preview
          rooftop:
            type: boolean
            description: whether or not the showing is on a rooftop
          dbox:
            type: boolean
            description: whether or not the showing is in D-Box seats
          dolbycinema:
            type: boolean
            description: whether or not the showing uses Dolby Cinema.
responses:
  UnauthorizedError:
    description: API key is missing or invalid
    headers:
      WWW_Authenticate:
        type: string
  BadRequestError:
    description: Bad Request. This reason is described in the body.
    schema:
      type: object
      required:
        - errors
      properties:
        errors:
          type: array
          items:
            title: Error Message
            description: The reason for creation failure.
            type: string
  InternalServerError:
    description: Internal server error
    schema:
      type: object
      required:
        - message
      properties:
        message:
          title: Error Message
          description: The reason for server error
          type: string